% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlagstaffMaxTemp.R
\docType{data}
\name{FlagstaffMaxTemp}
\alias{FlagstaffMaxTemp}
\title{A time series of daily maximum temperatures in Flagstaff, AZ.}
\format{
a data frame with 10882 observations
\describe{
  \item{Date}{The date of observation as a POSIX date format.}
  \item{MaxTemp}{Daily maximum temperature in degrees Farhenheit.}
}
}
\source{
\url{www.ncdc.noaa.gov}
}
\usage{
FlagstaffMaxTemp
}
\description{
A time series of daily maximum temperatures in Flagstaff, AZ.
}
\keyword{datasets}
